Refactor storage to separate SQLite-specific logic.
Introduce a dedicated `SqliteStore` for better modularization and clarity of SQLite database operations. Renamed and reorganized files to reflect the new structure, aligning SQLite-specific functionality under appropriate modules.
